+++ src/modules/FixIgnoreDamage.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "../Module.hpp"
+#include <Geode/modify/PlayerObject.hpp>
+
+class $editorModule(FixIgnoreDamage) {
+    bool onToggled(bool state) override;
+};
+
+class $modify(FIDPlayerObject, PlayerObject) {
+    $registerEditorHooks(FixIgnoreDamage)
+
+    void collidedWithSlopeInternal(float dt, GameObject* object, bool forced);
+};